Fixed brackets

+++ src/position.c
@@ -17,7 +17,7 @@ void position_init()
 
 uint32_t get_pos()
 {
-  uint32_t t = (uint32_t) pos_timer_read();
+  uint32_t t = pos_timer_read();
   return (offset + (t * POS_IN_A_TURN / countPerTour)) % POS_IN_A_TURN;
 }
